refactor(calc): Change package name
