+++ app/src/main/java/com/example/yosuke/calculator/model/entity/CalcEntity.kt
@@ -1,4 +1,4 @@
-package com.example.yosuke.calculator.model.elements
+package com.example.yosuke.calculator.model.entity
 
 data class CalcEntity(val number: Long, val operator: Operators) {
     override fun toString(): String {
