+++ app/src/main/java/com/example/yosuke/calculator/model/usecase/CalcUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.yosuke.calculator.model.usecase
 
-import com.example.yosuke.calculator.model.elements.Operators
+import com.example.yosuke.calculator.model.entity.Operators
 
 interface CalcUseCase {
     fun calc(x: Long, operator: Operators, y: Long): Long
