+++ app/src/main/java/com/example/yosuke/calculator/model/usecase/CalcUseCaseImpl.kt
@@ -1,6 +1,6 @@
 package com.example.yosuke.calculator.model.usecase
 
-import com.example.yosuke.calculator.model.elements.Operators
+import com.example.yosuke.calculator.model.entity.Operators
 
 class CalcUseCaseImpl : CalcUseCase {
     override fun calc(x: Long, operator: Operators, y: Long): Long {
